feat: enhance subject comment reactions and refactor topic article module

+++ shared/data/src/commonMain/kotlin/com/xiaoyv/bangumi/shared/data/api/next/SubjectApi.kt
@@ -123,7 +123,7 @@ interface SubjectApi {
      */
     @GET("p1/subjects/{subjectID}/comments")
     suspend fun getSubjectComments(
-        @Path("subjectID") subjectID: Int,
+        @Path("subjectID") subjectID: Long,
         @Query("type") type: CollectionType? = null,
         @Query("limit") limit: Int? = 20,
         @Query("offset") offset: Int? = 0,
